// src/types/regions.ts
export type Province = {
  id: number
  name: string
}

export type Regency = {
  id: number
  name: string
  province_id: number
}

export type District = {
  id: number
  name: string
  regency_id: number
}

export type RegionsPayload = {
  provinces: Province[]
  regencies: Regency[]
  districts: District[]
}

export type RegionFilters = {
  provinceId: number | null
  regencyId: number | null
  districtId: number | null
}